ENH: Test brainmatch score computing script
Test brainmatch score computing script.

Avoid providing the `brainmatch` package name to `pytest` in the workflow so
that the score computing script gets tested.

Allow the rest of the files in the `data` folder to be retrieved in tests.

Add the necessary `pytest_console_scripts` package to the dev requirements.